test_parse/pyproject.toml
[tool.poetry]
name = "test-parse"
version = "0.1.0"
description = ""
authors = ["Mykola322 <mikolasutij2009@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
